Refactor info collect : divide collect and print in 3 files
These file concern :
- Info collecting
- file content
- file writing
