+++ src/libs/antares/CMakeLists.txt
@@ -342,8 +342,12 @@ set(SRC_BENCHMARKING
     benchmarking/i_writer.h
     benchmarking/csv_writer.h
     benchmarking/csv_writer.cpp
-    benchmarking/studyinfo.h
-    benchmarking/studyinfo.cpp
+    benchmarking/info_collectors.h
+    benchmarking/info_collectors.cpp
+	benchmarking/file_content.h
+    benchmarking/file_content.cpp
+	benchmarking/file_writer.h
+	benchmarking/file_writer.cpp
 	)
 source_group("misc\\benchmarking" FILES ${SRC_BENCHMARKING})
