+++ src/libs/antares/benchmarking.h
@@ -30,6 +30,7 @@
 #include "benchmarking/timer.h"
 #include "benchmarking/content_handler.h"
 #include "benchmarking/csv_writer.h"
-#include "benchmarking/studyinfo.h"
+#include "benchmarking/info_collectors.h"
+#include "benchmarking/file_writer.h"
 
 #endif // __ANTARES_LIBS_BENCHMARKING_H__
