+++ src/libs/antares/benchmarking/file_writer.h
@@ -1,0 +1,32 @@
+#pragma once
+
+#include <yuni/yuni.h>
+#include <yuni/io/file/stream.h>
+
+#include "file_content.h"
+
+namespace Benchmarking
+{
+class FileWriter
+{
+public:
+    FileWriter(FileContent& fileContent);
+    virtual void flush() = 0;
+
+protected:
+    // Member data
+    FileContent& fileContent_;
+};
+
+class FileCSVwriter final : public FileWriter
+{
+public:
+    explicit FileCSVwriter(Yuni::String& filePath, FileContent& fileContent);
+    void flush() override;
+private:
+    Yuni::IO::File::Stream outputFile_;
+    Yuni::String& filePath_;
+};
+
+} // namespace Benchmarking
+
